requirements.txt
requests
pandas
websocket-client
streamlit